# servers/shutter/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Get Shutter List","operationId":"get_shutter_list__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShutterChannelList"}}}}}}},"/token":{"post":{"summary":"Login For Access Token","operationId":"login_for_access_token_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{shutter_name}":{"get":{"summary":"Get Shutter State","operationId":"get_shutter_state__shutter_name__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"shutter_name","in":"path","required":true,"schema":{"type":"string","title":"Shutter Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShutterStateReport"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Set Shutter State","operationId":"set_shutter_state__shutter_name__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"shutter_name","in":"path","required":true,"schema":{"type":"string","title":"Shutter Name"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShutterChannelOperation"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShutterStateReport"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{shutter_name}/ws":{"get":{"summary":"Ws Endpoint","operationId":"ws_endpoint__shutter_name__ws_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"shutter_name","in":"path","required":true,"schema":{"type":"string","title":"Shutter Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_for_access_token_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_token_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ShutterAction":{"type":"string","enum":["ON","OFF","SWITCH"],"title":"ShutterAction"},"ShutterChannelList":{"properties":{"shutter_list":{"items":{"type":"string"},"type":"array","title":"Shutter List"}},"type":"object","required":["shutter_list"],"title":"ShutterChannelList"},"ShutterChannelOperation":{"properties":{"action":{"$ref":"#/components/schemas/ShutterAction"}},"type":"object","required":["action"],"title":"ShutterChannelOperation"},"ShutterState":{"type":"string","enum":["OFF","ON"],"title":"ShutterState"},"ShutterStateReport":{"properties":{"shutter_name":{"type":"string","title":"Shutter Name"},"state":{"$ref":"#/components/schemas/ShutterState"}},"type":"object","required":["shutter_name","state"],"title":"ShutterStateReport"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}